+++ projects/tokens/contracts/beacons/ERC1155RaribleBeacon.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.7.0;
 import "@openzeppelin/contracts/proxy/UpgradeableBeacon.sol";
 
 contract ERC1155RaribleBeacon is UpgradeableBeacon {
-    constructor(address impl) UpgradeableBeacon(impl) {
-
+    constructor(address impl, address initialOwner) UpgradeableBeacon(impl) {
+        transferOwnership(initialOwner);
     }
 }