+++ projects/tokens/test-rarible-user-1155.sh
@@ -2,4 +2,4 @@
 truffle test ./test/erc-1155/ERC1155RaribleUser.test.js \
               ./contracts/create-2/ERC1155RaribleFactoryC2.sol \
               ./test/contracts/TestERC1271.sol \
-              ./test/contracts/TestRoyaltyV2981Calculate.sol
+              ./test/contracts/TestRoyaltyV2981Calculate.sol --compile-all