+++ projects/transfer-manager/test-lib-fee-side.sh
@@ -2,7 +2,7 @@
 #Test doTransfers in exchange-v2/contracts/exchange/v2/RaribleTransferManager.sol
 truffle test \
             ./test/LibFeeSide.test.js \
-            ./test/contracts/LibFeeSideTest.sol
+            ./test/contracts/LibFeeSideTest.sol --compile-all