+++ projects/tokens/CHANGELOG.md
@@ -38,7 +38,26 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - add missing deps ([243f743](https://github.com/rarible/protocol-contracts/commit/243f7436750c401c7b1dbe44ce9a6de1ca04f7f4))
 - add new config ([#289](https://github.com/rarible/protocol-contracts/issues/289)) ([266cf8c](https://github.com/rarible/protocol-contracts/commit/266cf8cc237d009dbd98850120f54ac10ef66a8e))
+- refactoring ([#283](https://github.com/rarible/protocol-contracts/issues/283)) ([014b6a3](https://github.com/rarible/protocol-contracts/commit/014b6a3ea4331c7334a0b5299614eb56fc5d58f3))
+- sanity check script for after new chain deployment ([#329](https://github.com/rarible/protocol-contracts/issues/329)) ([f046cc5](https://github.com/rarible/protocol-contracts/commit/f046cc564fa613894284c50b71eba44ed06e4483))
+
+**Note:** Version bump only for package @rarible/tokens
+
+## [0.8.6](https://github.com/rarible/protocol-contracts/compare/v0.8.1...v0.8.6) (2024-04-27)
+
+### Features
+
 - PT-5072 add seaport 1.6 to exchange wrapper, deploy wrapper to sepolia ([c6f10f1](https://github.com/rarible/protocol-contracts/commit/c6f10f151740dd76da0e997b8bf499f03cba00a9))
+
+## [0.8.5](https://github.com/rarible/protocol-contracts/compare/v0.8.1...v0.8.5) (2024-04-27)
+
+### Features
+
+- add proxy deploy support ([40270b0](https://github.com/rarible/protocol-contracts/commit/40270b07cdb2e061a3a349535f7bb9717e686eb7))
+- change token factories back, optomise ERC1155 size ([8e25ae2](https://github.com/rarible/protocol-contracts/commit/8e25ae2ae5c03166ef3524935c827f168f6d0445))
+- fix hardhat-migrations, change createToken() and getAddress() in token factories ([8930c47](https://github.com/rarible/protocol-contracts/commit/8930c47eec9da0b3bcbe5c6a467cae3e803085e2))
+- PT-5072 add seaport 1.6 to exchange wrapper, deploy wrapper to sepolia ([c6f10f1](https://github.com/rarible/protocol-contracts/commit/c6f10f151740dd76da0e997b8bf499f03cba00a9))
+- rename ([a2c6a83](https://github.com/rarible/protocol-contracts/commit/a2c6a83b94a89fc9aaddfa1e754cbf5c1280b8c7))
 
 ## [0.8.4-beta.3](https://github.com/rarible/protocol-contracts/compare/v0.8.4-beta.2...v0.8.4-beta.3) (2024-01-19)